+++ src/pages/Restaurants.vue
@@ -31,7 +31,31 @@
 
      
     </section>
+<!-- seleziona ristoranti  -->
+<section>
+    <button @click="showAllTypes"> <i class="fa-solid fa-magnifying-glass"></i>cerca tipologia</button>
+    <div v-if="showTypes">
+        <label for="types"></label>
+        <select name="types" id="types"  v-model="selectedType">
+            <option v-for="type in types" :value="type.id" @click="filterRestaurants">
+                {{ type.name }}
+            </option>
+        </select>
+        <!-- <button >Filtra ristoranti</button> -->
+    </div>
+</section>
 
+<section v-if="selectedType">
+    <div v-if="filteredRestaurants.length">
+        <h2>Ristoranti appartenenti alla tipologia: {{ selectedType }}</h2>
+        <div v-for="restaurant in filteredRestaurants" :key="restaurant.id" >
+            {{ restaurant.name }}
+        </div>
+    </div>
+    <div v-else>
+        <p>al momento non ci sono ristornati per questa tipologia</p>
+    </div>
+</section>
    
 
 
@@ -55,8 +79,11 @@ export default {
         return {
             store,
             restaurants: [],
-            
-           
+            types: [],
+            showTypes: false,
+            selectedType: null,
+            filteredRestaurants: []
+
         }
     },
 
@@ -80,6 +107,19 @@ export default {
                     console.log(response.data.results)
                     this.plates = response.data.results
                 })
+            },
+            showAllTypes() {
+                this.showTypes = !this.showTypes;
+            },
+            filterRestaurants(){
+                const data = {params: {
+                        type_id: this.selectedType, 
+                    }};
+                axios.get("http://127.0.0.1:8000/api/filter", data).then((response) => {
+                console.log(response.data.results);
+                this.filteredRestaurants = response.data.results;
+                console.log(this.filteredRestaurants);
+            });
             }
 
         },
@@ -87,7 +127,10 @@ export default {
         mounted(){
             this.getRestautants()
             this.getPlate()
-       
+            axios.get("http://127.0.0.1:8000/api/types").then((response) => {
+                // console.log(response.data.results);
+                this.types = response.data.results;
+            });
         }
     }
